Try to fix Formik performance
Formik seems to rerender the whole component on every keypress.  Since
Formik is relatively high up our component tree (in order to be able to
    render its buttons into the AppBar) those rerenders are noticeably
time consuming. Some issues on Github recommend using <FastField>, but
this did not seem to prevent any rerenders in our case. Hopefully this
issue will be resolved with React Suspense and a new version of Formik.