Edit subforms in dialog - direct mutation
This commit is not entirely cleaned up but shows the basic procedure of
editing parts of a formik form in a dialog. However, the fields in the
dialog mutate the form state directly, so the form cannot have any
reasonable confirm/cancel action. This is inconsistent UX in comparison
to the other dialogs in the application, but after some discussion we
concluded that it's probably not that big of an issue.

A version where the dialog kept its own state and commited them to the
parent form with a "Save" button would be constistent, but I'm not sure
whether that's a better alternative.