Clean up form value casting
Before we weren't able to pass the data from the backend directly into
an edit form, because fields that are null are not valid values for
React input components. Our base `InputFieldWithValidation` now
automatically passes "" instead of `null` to the input field.

For sending the values properly formatted to the backend, now transforms
with yup are used instead of a `cast` method in the store.