Fixed creation of Relationships in Factories
Previously, the Factories created too many objects (e.g., you wanted to seed a ProjectPosition and you already gave a project_id as an argument in the create function, it would still create a new project regardless). Now all the relationship calls are wrapped into closures, which won't be executed if you provide a relationship in your make or create arguments.