Added Save and Continue Button to TimetrackFormDialog
Basically, the Dialog won't close if you click this option. I also changed the behaviour of effortStore.doPost to not store the result from the post in the entity variable, so the Dialog will always use the effortTemplate observable, which contains no id and so will always trigger the post instead of put. This change does not apply to the put, that one does still overwrite the observable entity from the effortStore.