Fixed issue with TimeEffortValueField
If multiple Rate Units with the same factor were available, it always took the first one, even if the user selected another one. This was mainly because I did not think about the fact that maybe there could be archived Rate Units with same factors. Now it separatly tracks the factor and the ID, so the correct RateUnit will be displayed.