Semaphore-Konfiguration für PHP-Dime (#124)
* Added pipeline configuration for Semaphore

* renamed web client to frontend

* extracted check backend formatting for semaphore

* Added test_backend command to api container

* Add wait-for-it to prevent failures

If the db container is not booted up finaly before running the migration command it might fail. So we wait-for-it

* Add missing wget

* Fixed test for budget price calculation

I'm a bit confused why this test worked before. The CostBreakdown also calculates VAT and the test previously ignored that. Anyway, now the creation of the objects also includes a fixed VAT and the expected result is updated.

* Fixed test for effective time of Work Period

This date has to be in the future, so it returns 0. Now the date is so far in the future, it will survive even the betterBetterBetterBetterBetterDime