allow deletion of project_position even if it has invoice_positions.
This allows you to modify the services of a project without having to deal with old invoices that might be based on these services. As far as I can tell, the invoice position <--> project position relation is not visible in the frontend anyways and is only used for generating reports for an invoice?